requirements.txt
streamlit
spotipy
pandas
numpy
scikit-learn
python-dotenv
tenacity
openai
